# benefit-contract/src/main/resources/openapi/benefit-center-v1.yaml
openapi: 3.0.3
info:
  title: Enterprise Benefit Center API
  version: 1.0.0
  description: |
    Company-wide idempotent benefit fulfillment API. Tenant identity is resolved from a verified JWT;
    callers cannot select a tenant or fulfillment channel in the body. V1 package policy is BEST_EFFORT.
servers:
  - url: https://benefit.example.internal
security:
  - bearerAuth: []
...
paths:
  /openapi/v1/award-orders:
    post:
      tags: [Awards]
      operationId: acceptAwardIntent
      summary: Idempotently accept an award intent
      description: Idempotency-Key must equal sourceRequestId in v1. A replay returns the original order.
      security:
        - bearerAuth: [benefit.award.write]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AwardIntent' }
      responses:
        '202':
          description: Accepted, including an idempotent replay
          headers:
            Location: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AcceptResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
    get:
      tags: [Awards]
      operationId: findAwardOrderBySource
      summary: Recover an order by the producer idempotency identity
      security:
        - bearerAuth: [benefit.award.read]
      parameters:
        - in: query
          name: sourceSystem
          required: true
          schema: { type: string, maxLength: 64 }
        - in: query
          name: sourceRequestId
          required: true
          schema: { type: string, maxLength: 128 }
      responses:
        '200':
          description: Found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AwardOrder' }
        '404': { $ref: '#/components/responses/NotFound' }
  /openapi/v1/award-orders/{awardOrderNo}:
    get:
      tags: [Awards]
      operationId: getAwardOrder
      summary: Query an award order and its item-level outcomes
      security:
        - bearerAuth: [benefit.award.read]
      parameters:
        - in: path
          name: awardOrderNo
          required: true
          schema: { type: string, maxLength: 64 }
      responses:
        '200':
          description: Found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AwardOrder' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /internal/v1/remediations:
    post:
      tags: [Remediation]
      operationId: acceptRemediation
      summary: Propose an audited reissue, reversal, or manual review
      description: UNKNOWN original operations are rejected. Non-manual actions require approvalRef.
      security:
        - bearerAuth: [benefit.remediate]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RemediationCommand' }
      responses:
        '200':
          description: Proposed or approved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RemediationResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /internal/v1/remediations/{remediationNo}:
    get:
      tags: [Remediation]
      operationId: getRemediation
      security:
        - bearerAuth: [benefit.remediate]
      parameters:
        - $ref: '#/components/parameters/RemediationNo'
      responses:
        '200':
          description: Current remediation state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RemediationResult' }
        '404': { $ref: '#/components/responses/NotFound' }
  /internal/v1/remediations/{remediationNo}/execute:
    post:
      tags: [Remediation]
      operationId: executeRemediation
      security:
        - bearerAuth: [benefit.remediate]
      parameters:
        - $ref: '#/components/parameters/RemediationNo'
        - $ref: '#/components/parameters/IdempotencyKey'
        - in: header
          name: X-Worker-Id
          schema: { type: string, maxLength: 128 }
      responses:
        '200':
          description: Dispatch state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RemediationResult' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /admin/v1/inventory/adjustments:
    post:
      tags: [Administration]
      operationId: adjustCenterInventory
      security:
        - bearerAuth: [benefit.admin]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InventoryCommand' }
      responses:
        '202': { description: Adjustment appended to the inventory ledger }
  /admin/v1/console/me:
    get:
      tags: [Console]
      operationId: getConsoleIdentity
      summary: Current console identity for the mapped JWT tenant
      security:
        - bearerAuth: [benefit.admin]
      responses:
        '200':
          description: Identity
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConsoleIdentity' }
  /admin/v1/console/overview:
    get:
      tags: [Console]
      operationId: getConsoleOverview
      summary: Tenant-scoped operational counts for the console dashboard
      security:
        - bearerAuth: [benefit.admin]
      responses:
        '200':
          description: Counts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConsoleOverview' }
  /admin/v1/console/attention-orders:
    get:
      tags: [Console]
      operationId: listAttentionOrders
      summary: Orders that are partial, remediating, or contain UNKNOWN/QUERYING items
      security:
        - bearerAuth: [benefit.admin]
      parameters:
        - $ref: '#/components/parameters/ConsoleLimit'
      responses:
        '200':
          description: Attention list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AttentionOrder' }
  /admin/v1/tenants/current:
    get:
      tags: [Console]
      operationId: getCurrentTenant
      security:
        - bearerAuth: [benefit.admin]
      responses:
        '200':
          description: Current tenant config
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TenantView' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/v1/skus:
    get:
      tags: [Console]
      operationId: listBenefitSkus
      security:
        - bearerAuth: [benefit.admin]
      parameters:
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/SkuTemplateStatus' }
        - in: query
          name: afterSkuId
          description: Seek cursor; return rows whose skuId is greater than this value.
          schema: { type: string, maxLength: 128 }
        - $ref: '#/components/parameters/ConsoleLimit'
      responses:
        '200':
          description: SKU list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SkuView' }
  /admin/v1/wallets/{subjectRef}:
    get:
      tags: [Console]
      operationId: getWallet
      summary: Get a tenant-scoped customer wallet summary
      security:
        - bearerAuth: [benefit.admin]
      parameters:
        - $ref: '#/components/parameters/SubjectRef'
      responses:
        '200':
          description: Wallet summary
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WalletView' }
  /admin/v1/wallets/{subjectRef}/entries:
    get:
      tags: [Console]
      operationId: listWalletEntries
      summary: List wallet entries with seek pagination
      security:
        - bearerAuth: [benefit.admin]
      parameters:
        - $ref: '#/components/parameters/SubjectRef'
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/WalletEntryStatus' }
        - in: query
          name: skuId
          schema: { type: string, maxLength: 128 }
        - in: query
          name: afterEntryId
          description: Seek cursor; return rows whose entryId is greater than this value.
          schema: { type: string, maxLength: 64 }
        - $ref: '#/components/parameters/ConsoleLimit'
      responses:
        '200':
          description: Wallet entries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/WalletEntryView' }
  /admin/v1/routes:
    get:
      tags: [Console]
      operationId: listChannelRoutes
      security:
        - bearerAuth: [benefit.admin]
      parameters:
        - in: query
          name: skuId
          schema: { type: string, maxLength: 128 }
        - $ref: '#/components/parameters/ConsoleLimit'
      responses:
        '200':
          description: Route list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/RouteView' }
  /admin/v1/inventory/accounts:
    get:
      tags: [Console]
      operationId: listInventoryAccounts
      security:
        - bearerAuth: [benefit.admin]
      parameters:
        - $ref: '#/components/parameters/ConsoleLimit'
      responses:
        '200':
          description: Inventory balances
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/InventoryView' }
  /admin/v1/remediations:
    get:
      tags: [Console]
      operationId: listRemediations
      security:
        - bearerAuth: [benefit.admin]
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - $ref: '#/components/parameters/ConsoleLimit'
      responses:
        '200':
          description: Remediation list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/RemediationView' }
  /admin/v1/code-assets:
    get:
      tags: [Console]
      operationId: listCodeAssets
      description: Returns metadata only. cipherText is never included; codeHash is truncated to 8 characters.
      security:
        - bearerAuth: [benefit.admin]
      parameters:
        - in: query
          name: skuId
          schema: { type: string, maxLength: 128 }
        - $ref: '#/components/parameters/ConsoleLimit'
      responses:
        '200':
          description: Code asset metadata
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CodeAssetView' }
    post:
      tags: [Administration]
      operationId: importEncryptedCodeAsset
      security:
        - bearerAuth: [benefit.admin]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CodeAssetCommand' }
      responses:
        '202': { description: Imported; plaintext codes are not accepted by this contract }
...
components:
  securitySchemes:
    bearerAuth:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://auth.example.internal/oauth/token
          scopes:
            benefit.award.write: Submit award intents
            benefit.award.read: Query award state
            benefit.remediate: Submit and query remediation
            benefit.admin: Manage catalog and owned inventory
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema: { type: string, minLength: 1, maxLength: 128 }
    RemediationNo:
      in: path
      name: remediationNo
      required: true
      schema: { type: string, maxLength: 64 }
...
    ConsoleLimit:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
    SubjectRef:
      in: path
      name: subjectRef
      required: true
      schema: { type: string, minLength: 1, maxLength: 256 }
  responses:
    BadRequest:
      description: Invalid intent or command
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: Missing or invalid authentication
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: Missing scope, invalid tenant mapping, or cross-tenant envelope
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: Idempotency payload conflict or invalid state transition
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: Resource not found in the authenticated tenant
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    AwardIntent:
      type: object
      additionalProperties: false
      required: [schemaVersion, sourceSystem, sourceRequestId, recipientRef, partialPolicy, items]
      properties:
        schemaVersion: { type: string, enum: ['1.0'] }
        sourceSystem: { type: string, minLength: 1, maxLength: 64 }
        sourceRequestId: { type: string, minLength: 1, maxLength: 128 }
        sourceBusinessNo: { type: string, maxLength: 128, nullable: true }
        recipientRef: { type: string, minLength: 1, maxLength: 256 }
        decision: { $ref: '#/components/schemas/DecisionReference' }
        partialPolicy: { type: string, enum: [BEST_EFFORT] }
        items:
          type: array
          minItems: 1
          maxItems: 20
          items: { $ref: '#/components/schemas/AwardItemIntent' }
        trace:
          type: object
          additionalProperties: { type: string, maxLength: 256 }
    DecisionReference:
      type: object
      additionalProperties: false
      properties:
        decisionId: { type: string, maxLength: 128 }
        activityId: { type: string, maxLength: 64 }
        activityVersion: { type: integer, format: int32, minimum: 1 }
    AwardItemIntent:
      type: object
      additionalProperties: false
      required: [clientItemId, benefitSkuId, benefitType, quantity]
      properties:
        clientItemId: { type: string, minLength: 1, maxLength: 128 }
        benefitSkuId: { type: string, minLength: 1, maxLength: 128 }
        benefitType: { $ref: '#/components/schemas/BenefitType' }
        amountMinor: { type: integer, format: int64, minimum: 1, nullable: true }
        currency: { type: string, pattern: '^[A-Z]{3}$', nullable: true }
        quantity: { type: integer, format: int64, minimum: 1, maximum: 1 }
        metadata:
          type: object
          additionalProperties: { type: string, maxLength: 256 }
      description: Atomic item (quantity=1). Split multiple units into stable clientItemIds. CASH requires amountMinor/currency; every non-cash type forbids both monetary fields.
    BenefitType:
      type: string
      enum: [CASH, COUPON, SERVICE_VOUCHER, REDEMPTION_CODE, PHYSICAL]
    AcceptResult:
      type: object
      required: [awardOrderNo, status, replay]
      properties:
        awardOrderNo: { type: string }
        status: { $ref: '#/components/schemas/AwardOrderStatus' }
        replay: { type: boolean }
    AwardOrder:
      type: object
      required: [orderNo, sourceSystem, sourceRequestId, recipientRef, status, homeCell, items]
      properties:
        orderNo: { type: string }
        sourceSystem: { type: string }
        sourceRequestId: { type: string }
        sourceBusinessNo: { type: string, nullable: true }
        recipientRef: { type: string }
        status: { $ref: '#/components/schemas/AwardOrderStatus' }
        homeCell: { type: string }
        items:
          type: array
          items: { $ref: '#/components/schemas/AwardItem' }
    AwardOrderStatus:
      type: string
      enum: [ACCEPTED, PROCESSING, SUCCEEDED, PARTIAL_SUCCEEDED, FAILED, REMEDIATING, REVERSING, REVERSED, PARTIALLY_REVERSED]
    AwardItem:
      type: object
      properties:
        itemNo: { type: string }
        clientItemId: { type: string }
        skuId: { type: string }
        benefitType: { $ref: '#/components/schemas/BenefitType' }
        quantity: { type: integer, format: int64 }
        amountMinor: { type: integer, format: int64, nullable: true }
        currency: { type: string, nullable: true }
        status:
          type: string
          enum: [PENDING, RESERVED, DISPATCHING, SUCCEEDED, FAILED_FINAL, UNKNOWN, QUERYING, REISSUING, REVERSING, REVERSED, REVERSAL_FAILED, REVERSAL_UNKNOWN]
        routeId: { type: string, nullable: true }
        failureCode: { type: string, nullable: true }
        latestOperationNo: { type: string, nullable: true }
        latestOperationStatus: { type: string, nullable: true }
        skuVersion: { type: integer, format: int64, minimum: 0 }
        walletEntryId: { type: string, nullable: true }
    RemediationCommand:
      type: object
      additionalProperties: false
      required: [externalCommandId, action, awardItemNo, reason]
      properties:
        externalCommandId: { type: string, minLength: 1, maxLength: 128 }
        action: { type: string, enum: [REISSUE, REVERSE, MANUAL_REVIEW] }
        awardItemNo: { type: string, minLength: 1, maxLength: 64 }
        originalOperationNo: { type: string, maxLength: 64, nullable: true }
        reason: { type: string, minLength: 1, maxLength: 512 }
        approvalRef: { type: string, maxLength: 256, nullable: true }
    RemediationResult:
      type: object
      properties:
        externalCommandId: { type: string }
        remediationNo: { type: string }
        status: { type: string, enum: [PROPOSED, APPROVED, REJECTED, DISPATCHING, SUCCEEDED, FAILED, UNKNOWN] }
        reference: { type: string, nullable: true }
        errorCode: { type: string, nullable: true }
...
    SkuTemplateStatus:
      type: string
      enum: [DRAFT, PENDING_APPROVAL, ACTIVE, PAUSED, RETIRED]
    ValidityType:
      type: string
      enum: [ABSOLUTE, RELATIVE]
...
    InventoryCommand:
      type: object
      required: [accountId, skuId, ownerType, ownerId, deltaAvailable, requestId]
      properties:
        accountId: { type: string }
        skuId: { type: string }
        ownerType: { type: string, enum: [CENTER_QUOTA, CHANNEL_SHADOW, CENTER_STOCK] }
        ownerId: { type: string }
        deltaAvailable: { type: integer, format: int64 }
        requestId: { type: string }
    CodeAssetCommand:
      type: object
      required: [codeAssetId, skuId, codeHash, cipherText, keyVersion]
      properties:
        codeAssetId: { type: string }
        skuId: { type: string }
        codeHash: { type: string, description: SHA-256 or keyed hash of the code }
        cipherText: { type: string, description: KMS-encrypted code; plaintext is forbidden }
        keyVersion: { type: string }
        expiresAt: { type: string, format: date-time, nullable: true }
    ConsoleIdentity:
      type: object
      required: [tenantId, subject, displayName, scopes]
      properties:
        tenantId: { type: string }
        subject: { type: string }
        displayName: { type: string }
        scopes:
          type: array
          items: { type: string }
    ConsoleOverview:
      type: object
      required: [unknownOps, partialOrders, pendingRemediations, skuCount, enabledRoutes, activeTemplateCount, walletIssued24h]
      properties:
        unknownOps: { type: integer, format: int64 }
        partialOrders: { type: integer, format: int64 }
        pendingRemediations: { type: integer, format: int64 }
        skuCount: { type: integer, format: int64 }
        enabledRoutes: { type: integer, format: int64 }
        activeTemplateCount: { type: integer, format: int64 }
        walletIssued24h: { type: integer, format: int64 }
    TenantView:
      type: object
      required: [tenantId, homeCell, enabled, version]
      properties:
        tenantId: { type: string }
        homeCell: { type: string }
        enabled: { type: boolean }
        version: { type: integer, format: int64 }
    SkuView:
      type: object
      required: [skuId, benefitType, status, enabled, validityType, usableWeekdays, version]
      properties:
        skuId: { type: string }
        benefitType: { $ref: '#/components/schemas/BenefitType' }
        faceValueMinor: { type: integer, format: int64, nullable: true }
        currency: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/SkuTemplateStatus' }
        enabled: { type: boolean }
        validityType: { $ref: '#/components/schemas/ValidityType' }
        validFrom: { type: string, format: date-time, nullable: true }
        validTo: { type: string, format: date-time, nullable: true }
        relativeDays: { type: integer, nullable: true }
        usableWeekdays:
          type: array
          items: { type: integer }
        dailyQuota: { type: integer, format: int64, nullable: true }
        userLimitPerDay: { type: integer, format: int64, nullable: true }
        userLimitTotal: { type: integer, format: int64, nullable: true }
        equivalentSkuId: { type: string, nullable: true }
        approvalProcessDefinitionKey:
          type: string
          nullable: true
          description: Present only while status is PENDING_APPROVAL; value is benefitSkuGoLive.
        approvalBusinessKey:
          type: string
          nullable: true
          description: Present only while status is PENDING_APPROVAL; value is skuId.
        version: { type: integer, format: int64 }
    WalletEntryStatus:
      type: string
      enum: [UNUSED, FROZEN, USED, EXPIRED, REVERSED]
...
    WalletView:
      type: object
      required: [subjectRef, totalEntries, unusedEntries, cashBalances]
      properties:
        subjectRef: { type: string }
        totalEntries: { type: integer, format: int64 }
        unusedEntries: { type: integer, format: int64 }
        cashBalances:
          type: array
          items: { $ref: '#/components/schemas/CashBalanceView' }
    CashBalanceView:
      type: object
      required: [currency, balanceMinor]
      properties:
        currency: { type: string }
        balanceMinor: { type: integer, format: int64 }
    WalletEntryView:
      type: object
      required: [entryId, subjectRef, skuId, skuVersion, awardOrderNo, itemNo, assetType, status, version, createdAt]
      properties:
        entryId: { type: string }
        subjectRef: { type: string }
        skuId: { type: string }
        skuVersion: { type: integer, format: int64 }
        awardOrderNo: { type: string }
        itemNo: { type: string }
        assetType: { type: string, enum: [COUPON, CASH_BALANCE, CODE, PHYSICAL] }
        status: { $ref: '#/components/schemas/WalletEntryStatus' }
        version: { type: integer, format: int64, minimum: 0 }
        expiresAt: { type: string, format: date-time, nullable: true }
        faceValueMinor: { type: integer, format: int64, nullable: true }
        currency: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    RouteView:
      type: object
      required: [routeId, skuId, priority, channelCode, ownerType, reserveMode, enabled, version]
      properties:
        routeId: { type: string }
        skuId: { type: string }
        priority: { type: integer }
        channelCode: { type: string }
        ownerType: { type: string }
        fallbackRouteId: { type: string, nullable: true }
        reserveMode: { type: string }
        enabled: { type: boolean }
        configRef: { type: string, nullable: true }
        version: { type: integer, format: int64 }
    InventoryView:
      type: object
      required: [accountId, skuId, ownerType, ownerId, available, reserved, issued, version]
      properties:
        accountId: { type: string }
        skuId: { type: string }
        ownerType: { type: string }
        ownerId: { type: string }
        available: { type: integer, format: int64 }
        reserved: { type: integer, format: int64 }
        issued: { type: integer, format: int64 }
        version: { type: integer, format: int64 }
    AttentionOrder:
      type: object
      required: [orderNo, sourceSystem, sourceRequestId, status]
      properties:
        orderNo: { type: string }
        sourceSystem: { type: string }
        sourceRequestId: { type: string }
        status: { type: string }
        reason: { type: string }
        updatedAt: { type: string, format: date-time, nullable: true }
    RemediationView:
      type: object
      required: [remediationNo, action, itemNo, status]
      properties:
        remediationNo: { type: string }
        action: { type: string }
        itemNo: { type: string }
        status: { type: string }
        reason: { type: string, nullable: true }
        updatedAt: { type: string, format: date-time, nullable: true }
    CodeAssetView:
      type: object
      required: [codeAssetId, skuId, status]
      properties:
        codeAssetId: { type: string }
        skuId: { type: string }
        status: { type: string }
        codeHashPrefix: { type: string, nullable: true, description: First 8 characters of codeHash. Never the full hash or cipherText. }
        expiresAt: { type: string, format: date-time, nullable: true }
    Error:
      type: object
      description: RFC 9457 Problem Details with a stable business code.
      required: [type, title, status, detail, code, timestamp]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer, format: int32 }
        detail: { type: string }
        instance: { type: string, format: uri, nullable: true }
        code:
          type: string
          enum: [INVALID_INTENT, IDEMPOTENCY_PAYLOAD_CONFLICT, SKU_NOT_FOUND, SKU_NOT_ACTIVE, SKU_NOT_DRAFT, SKU_VERSION_CONFLICT, SKU_APPROVAL_LOCKED, SKU_ILLEGAL_TRANSITION, SKU_SUBMIT_IN_FLIGHT, WALLET_ENTRY_NOT_FOUND, WALLET_ILLEGAL_TRANSITION, WALLET_ALREADY_USED, WALLET_VERSION_CONFLICT, WALLET_EXPIRED, USER_LIMIT_EXCEEDED, INVENTORY_INSUFFICIENT, CHANNEL_RETRYABLE, CHANNEL_FINAL_FAILURE, CHANNEL_RESULT_UNKNOWN, REMEDIATION_NOT_ALLOWED, INTERNAL_ERROR]
        message: { type: string, deprecated: true, description: Compatibility alias of detail. }
        timestamp: { type: string, format: date-time }